calculate canGrow/canShrink based on tolerance
